feat(*): fixes de implementaciones anteriores, busqueda de ordenes por cliente

+++ src/main/java/com/progavanzada/taller/mecanico/controller/dto/MarcaCreateDto.java
@@ -14,5 +14,5 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 public class MarcaCreateDto {
     public String name;
     public String origen;
-    public float impuestoMarca;
+    public Integer impuestoMarca;
 }
